+++ android/app/src/main/java/io/coursepick/coursepick/presentation/createcustomcourse/CreateCustomCourseUiEvent.kt
@@ -1,0 +1,11 @@
+package io.coursepick.coursepick.presentation.createcustomcourse
+
+import io.coursepick.coursepick.domain.customcourse.DraftSegment
+
+sealed interface CreateCustomCourseUiEvent {
+    data class NewSegment(
+        val segment: DraftSegment,
+    ) : CreateCustomCourseUiEvent
+
+    data object RemoveLastWaypoint : CreateCustomCourseUiEvent
+}
