--- android/app/src/main/java/io/coursepick/coursepick/presentation/customcourse/CustomCourseUiEvent.kt
@@ -1,11 +1,0 @@
-package io.coursepick.coursepick.presentation.customcourse
-
-import io.coursepick.coursepick.domain.customcourse.DraftSegment
-
-sealed interface CustomCourseUiEvent {
-    data class NewSegment(
-        val segment: DraftSegment,
-    ) : CustomCourseUiEvent
-
-    data object RemoveLastWaypoint : CustomCourseUiEvent
-}
